+++ aegis_grpc/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
+* [PR-119](https://github.com/AGH-CEAI/aegis_ros/pull/119) - Server: parametrized topics subs buffer size (and changed default from `10` to `1`).
+
 ### Deprecated
 ### Removed
 ### Fixed